feat: enhance n8n nodes with TypeScript, OCR discovery, and security compliance
Major improvements to n8n community package:

🔧 TypeScript Migration:
- Convert all nodes from JavaScript to TypeScript
- Add proper TypeScript configuration (tsconfig.json)
- Update build system with TypeScript compilation
- Add type safety and better development experience

🔍 Enhanced Node Discovery:
- Add comprehensive OCR keywords to package.json
- Improve LLMWhisperer node description for better searchability
- Add dynamic subtitle showing current operation mode
- Create codex files (.node.json) for proper categorization

🛡️ Security Compliance:
- Replace setTimeout with Promise-based delay mechanism
- Ensure compliance with n8n community package security rules
- Remove restricted globals usage
- Pass all n8n security validation checks

📁 Improved Project Structure:
- Reorganize nodes into separate directories
- Add proper build tooling (gulp, prettier, eslint)
- Create comprehensive package configuration
- Add MIT license and proper documentation

✨ Node Enhancements:
- LLMWhisperer: Enhanced OCR functionality with better descriptions
- Unstract: Document processing with improved error handling
- UnstractHITL: Human-in-the-loop operations (Push/Fetch)
- All nodes: TypeScript types and better maintainability

🔨 Development Tools:
- Add linting with n8n-specific rules
- Configure prettier for code formatting
- Set up proper build pipeline
- Add comprehensive package metadata

The package now passes all n8n community security checks and is ready for publication and verification.
