+++ nodes/UnstractHitlPush/UnstractHitlPush.node.json
@@ -1,0 +1,20 @@
+{
+	"node": "n8n-nodes-base.unstractHitlPush",
+	"nodeVersion": "1.0",
+	"codexVersion": "1.0",
+	"categories": [
+		"Data & Storage"
+	],
+	"resources": {
+		"credentialDocumentation": [
+			{
+				"url": "https://docs.unstract.com/hitl/api"
+			}
+		],
+		"primaryDocumentation": [
+			{
+				"url": "https://docs.unstract.com/hitl"
+			}
+		]
+	}
+}
